requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.1
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
paramiko==3.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyNaCl==1.5.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.11.0
